+++ Changelog.md
@@ -12,7 +12,7 @@
 * Do not generate paths like `/a/b/c/config/boot.rb/../../Gemfile` to require and open things, create a proper path instead.
 * Remove the hack for loading the entire lib folder with a proper solution. [#3809](https://github.com/diaspora/diaspora/issues/3750)
 * Update and refactor the default public view `public/default.html` [#3811](https://github.com/diaspora/diaspora/issues/3811)
-
+* Write unicorn stderr and stdout [#3785](https://github.com/diaspora/diaspora/pull/3785)
 
 ## Features